backend/requirements.txt
# Core Framework
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# Database
sqlalchemy>=2.0.30
psycopg2-binary>=2.9.9

# Agentic AI (AWS Bedrock)
langchain>=0.2.0
langchain-aws>=0.1.0
langchain-core>=0.2.0
langgraph>=0.2.0
boto3>=1.26.0

# Data Generation
pandas>=2.2.2
numpy>=1.26.4
faker>=25.2.0

# Config & Utils
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# WebSocket
websockets>=12.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-xdist>=3.5.0
pytest-asyncio>=0.23.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.1
httpx>=0.24.0
